Chore: (`vertix-bot`) - Move ui-utils from `vertix-base` utility functions for UI templates
`ui-utils.ts` file was added to the `vertix-bot` project. This file includes utility functions important for UI template manipulation such as `uiUtilsWrapAsTemplate`, `uiUtilsDynamicElementsRearrange` and `uiGenerateCustomIdHash`. The implementation accounts for dynamic rearranging of elements and generation of custom id hash value.
