+++ packages/vertix-bot/src/ui-v2/_general/invalid-channel-type/invalid-channel-type-embed.ts
@@ -1,6 +1,6 @@
 import { ChannelType, Colors } from "discord.js";
 
-import { uiUtilsWrapAsTemplate } from "@vertix.gg/base/src/utils/ui";
+import { uiUtilsWrapAsTemplate } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import { UIEmbedBase } from "@vertix.gg/bot/src/ui-v2/_base/ui-embed-base";
