+++ packages/vertix-bot/src/ui-v2/channel-name-template/channel-name-template-embed.ts
@@ -1,6 +1,6 @@
-import { uiUtilsWrapAsTemplate } from "@vertix.gg/base/src/utils/ui";
-
 import { DEFAULT_DYNAMIC_CHANNEL_NAME_TEMPLATE } from "@vertix.gg/base/src/definitions/master-channel-defaults";
+
+import { uiUtilsWrapAsTemplate } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import { UIEmbedBase } from "@vertix.gg/bot/src/ui-v2/_base/ui-embed-base";
