+++ packages/vertix-bot/src/ui-v2/claim/start/claim-start-embed.ts
@@ -1,4 +1,4 @@
-import { uiUtilsWrapAsTemplate } from "@vertix.gg/base/src/utils/ui";
+import { uiUtilsWrapAsTemplate } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import { UIEmbedBase } from "@vertix.gg/bot/src/ui-v2/_base/ui-embed-base";
