+++ packages/vertix-bot/src/ui-v2/config-extras/config-extras-select-menu.ts
@@ -1,5 +1,5 @@
 
-import { uiUtilsWrapAsTemplate } from "@vertix.gg/base/src/utils/ui";
+import { uiUtilsWrapAsTemplate } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import { UI_GENERIC_SEPARATOR, UIInstancesTypes } from "@vertix.gg/bot/src/ui-v2/_base/ui-definitions";
 import { UIElementStringSelectMenu } from "@vertix.gg/bot/src/ui-v2/_base/elements/ui-element-string-select-menu";
