+++ packages/vertix-bot/src/ui-v2/dynamic-channel/dynamic-channel-component.ts
@@ -1,4 +1,4 @@
-import { uiUtilsDynamicElementsRearrange } from "@vertix.gg/base/src/utils/ui";
+import { uiUtilsDynamicElementsRearrange } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import { UIComponentBase } from "@vertix.gg/bot/src/ui-v2/_base/ui-component-base";
