+++ packages/vertix-bot/src/ui-v2/dynamic-channel/meta/clear-chat/embeds/dynamic-channel-meta-clear-chat-success-embed.ts
@@ -1,4 +1,4 @@
-import { uiUtilsWrapAsTemplate } from "@vertix.gg/base/src/utils/ui";
+import { uiUtilsWrapAsTemplate } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import { UIEmbedBase } from "@vertix.gg/bot/src/ui-v2/_base/ui-embed-base";
