+++ packages/vertix-bot/src/ui-v2/dynamic-channel/permissions/embeds/dynamic-channel-permissions-blocked-embed.ts
@@ -1,6 +1,6 @@
 import { Colors } from "discord.js";
 
-import { uiUtilsWrapAsTemplate } from "@vertix.gg/base/src/utils/ui";
+import { uiUtilsWrapAsTemplate } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import {
     DynamicChannelPermissionsAccessEmbed
