+++ packages/vertix-bot/src/ui-v2/dynamic-channel/permissions/embeds/dynamic-channel-permissions-granted-embed.ts
@@ -1,4 +1,4 @@
-import { uiUtilsWrapAsTemplate } from "@vertix.gg/base/src/utils/ui";
+import { uiUtilsWrapAsTemplate } from "@vertix.gg/bot/src/ui-v2/ui-utils";
 
 import {
     DynamicChannelPermissionsAccessEmbed
